performance optimization on or() and friends
Rather than use Result.match(), which yeilds a new object on every
invocation, the or, orElse, and, and andThen now have some new tricks up
their sleeves. If the operation is a no-op, the same object is returned
but cast to its new type. This avoids redundant unwrapping and
re-wrapping of objects.